requirements.txt
absl-py==0.9.0
appnope==0.1.0
arviz==0.10.0
astunparse==1.6.3
atari-py==0.2.6
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
box2d-py==2.3.8
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.14.0
cftime==1.2.1
chardet==3.0.4
cloudpickle==1.3.0
cvxopt==1.2.5
cycler==0.10.0
Cython==0.29.20
decorator==4.4.2
defusedxml==0.6.0
dunshire==0.1.1
entrypoints==0.3
fastprogress==1.0.0
future==0.18.2
gast==0.3.3
glfw==1.11.2
google-auth==1.18.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.14.1
grpcio==1.29.0
gym==0.17.2
h5py==2.10.0
idna==2.9
imageio==2.8.0
importlib-metadata==1.5.0
ipykernel==5.2.1
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.16.0
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.2
jupyterlab-server==1.1.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.1.0
lockfile==0.12.2
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mlrose==1.3.0
more-itertools==8.2.0
nashpy==0.0.19
nbconvert==5.6.1
nbformat==5.0.6
netCDF4==1.5.4
notebook==6.0.3
numpy==1.17.3
oauthlib==3.1.0
opencv-python==4.2.0.34
opt-einsum==3.2.1
packaging==20.1
pandas==0.25.3
pandoc==1.0.2
pandocfilters==1.4.2
parso==0.7.0
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
pluggy==0.13.1
ply==3.11
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.12.2
ptyprocess==0.6.0
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyglet==1.5.0
Pygments==2.6.1
pymc3==3.9.3
pymdptoolbox==4.0b3
pyparsing==2.4.4
pyrsistent==0.16.0
pystan==2.19.1.1
pytest==5.3.5
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
pyzmq==19.0.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.10.1
Send2Trash==1.5.0
six==1.13.0
sklearn==0.0
tensorboard==2.2.2
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
Theano==1.0.5
threadpoolctl==2.1.0
tornado==6.0.4
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.9
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1
xarray==0.16.1
zipp==2.1.0
